Increased correlation with deeper dictionary matrix and added multiplexed function.
Checked Peak Height - its calculating properly.

Scan rate has been implemented into AUC. It is precise but may not be accurate (values are about 1.9x the AUC I calculated on the CH).

Increased dictionary matrix allows the program to be more robust and requires less assumptions and input on the part of the user. More ability to analyze 'random' data with little to no assumptions.