Two methods for AUC calculation show precise measurements.
When comparing the area calculated for a shapely polygon vs. that calculated using a right hand reimann sum (using step potential as the width and the adjusted currents and linear baseline for height)  I get the same value for the AUC.

When compared to what I calculated on the CH, however, the values are different by around a factor of two.

With the precision of the two methods shown here I believe the CH is less accurate than the implemented methods.