Comprehensive clean up of the script
Updated TextFileExport to properly adjust the exported Txt file. Removed averages and std if electrode count is below 2. Removed Ratiometric data export is frequencies < 2.

Updated PostAnalysisDataExport to properly export data. Can change file location, handle, and contents. Adjusts with the new updates to TextFileExport

Updated axes creation to be more precise. Now up to standard of SOP for both FrequencyMap and ContinuousAnalysis

Updated CheckPoint to be able to properly check multipotentiostat files. Previously it did not check for the contents of the file, only that it existed. It now queries the file contents to find the total number of columns. If the list value of the chosen electrode is out of this range it will not be accepted and CheckPoint will not proceed.

Added 8 more electrodes to the ElectrodeListBox